+++ app/Helpers/Delete/Models/DeleteOrganisation.php
@@ -41,20 +41,25 @@ class DeleteOrganisation implements DeleteInterface
      */
     public function delete()
     {
-        $this->canDelete();
+        if (! $this->canDelete()) {
+            return $this->errorMessage;
+        }
+
         $this->deleteModels();
         $this->dissociateRelations();
         $this->deleteRelations();
         $this->customDeleteActions();
-        $this->organisation->delete();
+
+        if (count($this->errorMessage) === 0) {
+            $this->organisation->delete();
+        }
 
         return $this->errorMessage;
     }
 
-    /** Checks if the model can be deleted and sets error messages
-     */
-    public function canDelete()
+    public function canDelete(): bool
     {
+        return true;
     }
 
     /** Deletes models recursive
@@ -66,7 +71,7 @@ class DeleteOrganisation implements DeleteInterface
 //             * 20221031; Jos; Deze relatie bestaat niet meer?
 //             */
 //            $deleteQuotationRequest = new DeleteQuotationRequest($quotationRequest);
-//            $this->errorMessage = array_merge($this->errorMessage, $deleteQuotationRequest->delete());
+//            $this->errorMessage = array_merge($this->errorMessage, ( $deleteQuotationRequest->delete() ?? [] ) );
 //        }
     }
